consolidate ServiceAgentPlugin
