+++ packages/webpack/package.json
@@ -14,6 +14,7 @@
     "access": "public"
   },
   "dependencies": {
+    "@entangled/fetch": "^2.0.0-alpha.9",
     "ts-morph": "^10.0.2",
     "webpack": "^5.65.0",
     "webpack-virtual-modules": "^0.4.3"
